move Details.js and Filter.js into the container directory
